requirements.txt
openai
python-dotenv
numpy
pandas
matplotlib
notebook
streamlit